+++ packages/components/scripts/jobs/react/index.js
@@ -1,3 +1,4 @@
 export * from './createWrappers.js';
 export * from './esbuild.js';
+export * from './intrinsic-elements.js';
 export * from './tsc.js';